Remove db sync to prevent overwrites
Before this commit, the pad would replace its contents with the database's after the database received a pad content update. This was meant to keep the pad in sync.
There existed a small sliver of time between the update and the replacement where the user could enter text, which would be overwritten when the replacement finished. It also messed up cursor positions. I've decided to take this out. The database would be in "sync" anyway...